\name{censoredMatrix}
\alias{censoredMatrix}
\docType{data}
\title{
The double-censored data.
}
\description{
A data frame with 8 observations on the following 6 variables.
}
\format{
  \describe{
    \item{\code{L}}{a numeric vector}
    \item{\code{R}}{a numeric vector}
    \item{\code{C1}}{a numeric vector}
    \item{\code{C2}}{a numeric vector}
    \item{\code{C3}}{a numeric vector}
  }
}

\examples{
data(censoredMatrix)
str(censoredMatrix)
}
