\name{sccr-package}
\alias{sccr-package}
\alias{sccr}
\docType{package}
\title{
The Self-Consistent, Competing Risks (SC-CR) Algorithms}

\description{
This package describes an algorithm for producing fully non-parametric and self-consistent estimators of the cause-specific failure probabilities in the presence of interval-censoring and possible masking of the failure cause in a competing risks environment.   It is a generalization of Turnbull's (1976) classic univariate algorithm.  
The algorithm was published in Adamic et al. (2010) and Adamic & Caron (2014).
}
\author{
Peter Adamic, Alicja Wolny-Dominiak
Maintainer: <alicja.wolny-dominiak@ue.katowice.pl>
}
\references{
1. Adamic, P., Caron, S. (2014),
"SC-CR Algorithms with Informative Masking", 
Scandinavian Actuarial Journal, 2014(4), 339-351. \doi{10.1080/03461238.2012.693457} \cr

2. Adamic, P., Dixon, S., Gillis, D. (2010), 
"Multiple Decrement Modeling in the Presence of Interval
Censoring and Masking", Scandinavian Actuarial Journal, 2010(4), 312-327. \doi{10.1080/03461230903134780} \cr

3. Turnbull, B. (1976).  The Empirical Distribution Function with Arbitrarily Grouped, Censored and Truncated Data, 
Journal of the Royal Statistical Society. Series B (Methodological), 38(3), 290-295. \doi{10.1111/j.2517-6161.1976.tb01597.x}
}


