\name{censoredMatrix}
\alias{censoredMatrix}

\title{
The matrix of probabilities}

\description{
The matrix of probabilities corresponding double-censored data of the form: \cr
(0,3]; (2,5] (c = 1); [4,4] (c = 2); (2, inf), (6,7]; (3,inf) (c=3); \cr
(6,8] (c=1 or c=2 or c=3) [9,9] \cr
c - the number of couse = 1,2,3}

\examples{
data(censoredMatrix)
}
