\name{sccr-package}
\alias{sccr-package}
\alias{sccr}
\docType{package}
\title{
The Self-Consistent, Competing Risks (SC-CR) Algorithms}

\description{
The SC-SR Algorithm is used to calculate the cause-deleted life expectancy improvement for left and right censored data. In the version 2.0 the function creating the probability matrix from double-censored data is added.
}
\author{
Peter Adamic, Alicja Wolny-Dominiak
Maintainer: <alicja.wolny-dominiak@ue.katowice.pl>
}
\references{
1. Adamic, P., Caron, S. (2014),
"SC-CR Algorithms with Informative Masking", 
Scandinavian Actuarial Journal, 2014(4), 339-351. \cr

2. Adamic, P., Dixon, S., Gillis, D. (2010), 
"Multiple Decrement Modeling in the Presence of Interval
Censoring and Masking", Scandinavian Actuarial Journal, 2010(4), 312-327. \cr

3. Adamic, P., Ouadah, S. (2009), 
"A Kernel Method for Modeling Interval Censored Competing
Risks", South African Statistical Journal, 43(1), 1-20. \cr

4. Turnbull, B. (1976).  The Empirical Distribution Function with Arbitrarily Grouped, Censored and Truncated Data, 
Journal of the Royal Statistical Society. Series B (Methodological), 38(3), 290-295.
}


